change Track.last_chapter_read to Float
each TrackService can convert it to Int if decimal chapters are not supported

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/AnilistApi.kt
@@ -35,7 +35,7 @@ class AnilistApi(val client: OkHttpClient, interceptor: AnilistInterceptor) {
         return withContext(Dispatchers.IO) {
             val variables = jsonObject(
                 "mangaId" to track.media_id,
-                "progress" to track.last_chapter_read,
+                "progress" to track.last_chapter_read.toInt(),
                 "status" to track.toAnilistStatus()
             )
             createDate(track.started_reading_date)?.let {
@@ -68,7 +68,7 @@ class AnilistApi(val client: OkHttpClient, interceptor: AnilistInterceptor) {
         return withContext(Dispatchers.IO) {
             val variables = jsonObject(
                 "listId" to track.library_id,
-                "progress" to track.last_chapter_read,
+                "progress" to track.last_chapter_read.toInt(),
                 "status" to track.toAnilistStatus(),
                 "score" to track.score.toInt()
             )
